Add getTurnOrder and restructure Character
Restructured Character to be less cumbersome to create, also added
some functionality to display in a nicer way, a way to compare by
speeds which is used for getTurnOrder, and an Eq instance that
only compares speeds.

getTurnOrder is working as long as enemies are passed in before
friendlies. This should cause no problems as long as we remember
that order does matter.